// src/types/datos.ts
export type Datos = {
    game: string
    game_label: string
    developerLabel: string
    genres: string
    platforms: string
    firstPublication: string
    logo?: string
}

export const datosVacios: Datos = {
    game: "",
    game_label: "",
    developerLabel: "",
    genres: "",
    platforms: "",
    firstPublication: "",
    logo: ""
}